[locker] Move scanner capture widgets into their own file
Pure move of the capture page's chrome, shutter, mode toggle, thumbnail
and done button widgets into scanner_capture_widgets.dart so the page
file is left with the capture state machine.
